Added poor man's threading to the MonteCarlo calculation.
My guess is that the ran() function from NR is not thread safe (it is non-reentrant). So I simply duplicated Monte_SANSX to Monte_SANSX2, where each incarnation uses a different random number generator, either ran1() or ran3(). This means that currently only two processors are supported. Not a big deal. At least it works.



git-svn-id: http://danse.chem.utk.edu/svn/sans/XOP_Dev@458 8e55283c-c6c9-4f4c-b3dc-704b0a22af0c