Added wavelength distribution as part of the simulation
Fixed bug where pixel locations of (-1,0) row or column were incorrectly rounded towards zero, and ended up on the detector in row 0 or column 0, effectively doubling the count there.

Separated the 4 functions into 4 separate files, for ease in checking that they are all the same. Now the only difference in each file should be the RNG used. And they had better be different - I'm almost positive now that cross-calling of the same RNG by different threads is responsible for the crashing.



git-svn-id: http://danse.chem.utk.edu/svn/sans/XOP_Dev@711 8e55283c-c6c9-4f4c-b3dc-704b0a22af0c